backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
requests
python-dotenv
google-generativeai